Module Map.String

specalized when key type is string, more efficient than the generic type, its compare behavior is fixed using the built-in comparison

type key = string

module I : sig ... end

module N : sig ... end

module A : sig ... end

Belt.Array Utililites for Array functions

type 'a t = (key, 'a) N.t

val empty : 'a option

val isEmpty : 'a option -> bool

val singleton : 'a -> 'b -> ('a, 'b) N.node option

val minKey : ('a, 'b) N.node option -> 'a option

val minKeyUndefined : ('a, 'b) N.node option -> 'a option

val maxKey : ('a, 'b) N.node option -> 'a option

val maxKeyUndefined : ('a, 'b) N.node option -> 'a option

val minimum : ('a, 'b) N.node option -> ('a * 'b) option

val minUndefined : ('a, 'b) N.node option -> ('a * 'b) option

val maximum : ('a, 'b) N.node option -> ('a * 'b) option

val maxUndefined : ('a, 'b) N.node option -> ('a * 'b) option

val forEachU : ('a, 'b) N.node option -> ('a -> 'b -> unit) -> unit

val forEach : ('a, 'b) N.node option -> ('a -> 'b -> unit) -> unit

val mapU : ('a, 'b) N.node option -> ('b -> 'c) -> ('a, 'c) N.node option

val map : ('a, 'b) N.node option -> ('b -> 'c) -> ('a, 'c) N.node option

val mapWithKeyU : ('a, 'b) N.node option -> ('a -> 'b -> 'c) -> ('a, 'c) N.node option

val mapWithKey : ('a, 'b) N.node option -> ('a -> 'b -> 'c) -> ('a, 'c) N.node option

val reduceU : ('a, 'b) N.node option -> 'c -> ('c -> 'a -> 'b -> 'c) -> 'c

val reduce : ('a, 'b) N.node option -> 'c -> ('c -> 'a -> 'b -> 'c) -> 'c

val everyU : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> bool

val every : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> bool

val someU : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> bool

val some : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> bool

val keepU : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> ('a, 'b) N.node option

val keep : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> ('a, 'b) N.node option

val partitionU : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> ('a, 'b) N.node option * ('a, 'b) N.node option

val partition : ('a, 'b) N.node option -> ('a -> 'b -> bool) -> ('a, 'b) N.node option * ('a, 'b) N.node option

val size : ('a, 'b) N.node option -> int

val toList : ('a, 'b) N.node option -> ('a * 'b) list

val toArray : ('a, 'b) N.node option -> ('a * 'b) array

val keysToArray : ('a, 'b) N.node option -> 'a array

val valuesToArray : ('a, 'b) N.node option -> 'b array

val checkInvariantInternal : ('a, 'b) N.t -> unit

val set : (key, 'a) N.t -> key -> 'a -> (key, 'a) N.node option

val updateU : (key, 'a) N.t -> key -> ('a option -> 'a option) -> (key, 'a) N.t

val update : (key, 'a) N.t -> key -> ('a option -> 'a option) -> (key, 'a) N.t

val removeAux : (key, 'a) N.node -> key -> (key, 'a) N.t

val remove : (key, 'a) N.node option -> key -> (key, 'a) N.t

val removeMany0 : (key, 'a) N.node -> key A.t -> int -> int -> (key, 'a) N.t

val removeMany : (key, 'a) N.node option -> key A.t -> (key, 'a) N.t

val mergeMany : (key, 'a) N.t -> (key * 'a) A.t -> (key, 'a) N.t

val mergeArray : (key, 'a) N.t -> (key * 'a) A.t -> (key, 'a) N.t

val has : (I.key, 'a) I.N.t -> I.key -> bool

val cmpU : (I.key, 'a) I.N.node option -> (I.key, 'b) I.N.node option -> ('a -> 'b -> int) -> int

val cmp : (I.key, 'a) I.N.node option -> (I.key, 'b) I.N.node option -> ('a -> 'b -> int) -> int

val eqU : (I.key, 'a) I.N.node option -> (I.key, 'b) I.N.node option -> ('a -> 'b -> bool) -> bool

val eq : (I.key, 'a) I.N.node option -> (I.key, 'b) I.N.node option -> ('a -> 'b -> bool) -> bool

val get : (I.key, 'a) I.N.t -> I.key -> 'a option

val getUndefined : (I.key, 'a) I.N.t -> I.key -> 'a Js.Undefined.t

val getWithDefault : (I.key, 'a) I.N.t -> I.key -> 'a -> 'a

val getExn : (I.key, 'a) I.N.t -> I.key -> 'a

val split : I.key -> (I.key, 'a) I.N.node option -> 'a I.t * 'a option * 'a I.t

val mergeU : (I.key, 'a) I.N.t -> 'b I.t -> (I.key -> 'a option -> 'b option -> 'c option) -> (I.key, 'c) I.N.node option

val merge : (I.key, 'a) I.N.t -> 'b I.t -> (I.key -> 'a option -> 'b option -> 'c option) -> (I.key, 'c) I.N.node option

val fromArray : (I.key * 'a) array -> (I.key, 'a) I.N.node option